+++ previous-versions/files/src/codatfiles/sdkconfiguration.py
@@ -20,9 +20,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '3.0.0'
-    sdk_version: str = '0.3.0'
-    gen_version: str = '2.142.2'
-    user_agent: str = 'speakeasy-sdk/python 0.3.0 2.142.2 3.0.0 codat-files'
+    sdk_version: str = '0.3.1'
+    gen_version: str = '2.143.2'
+    user_agent: str = 'speakeasy-sdk/python 0.3.1 2.143.2 3.0.0 codat-files'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: