+++ previous-versions/files/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [python v0.3.0] previous-versions/files
 ### Releases
 - [PyPI v0.3.0] https://pypi.org/project/codat-files/0.3.0 - previous-versions/files
+
+## 2023-10-02 00:01:06
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Files.yaml
+- Speakeasy CLI 1.92.3 (2.143.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.3.1] previous-versions/files
+### Releases
+- [PyPI v0.3.1] https://pypi.org/project/codat-files/0.3.1 - previous-versions/files